services/simple/notification/requirements.txt
Flask==3.0.1
sendgrid==6.11
pika==1.3.2
python-dotenv==1.0.0
